+++ Graphics/objects/particles.py
@@ -1,6 +1,6 @@
 from ...Core.base import Vec1, Vec2, Color3, Color4, drawMode, System
 from ...Core.modules import random
-from ...Utils.frametimer import frameTimer
+from ...Utils.General.frametimer import frameTimer
 from ..objects.rectangle import Rectangle
 from .randomcolor import randomColor4