+++ Utils/System/manager.py
@@ -1,0 +1,6 @@
+class Manager:
+    def _find(self, name):
+        for index, obj in enumerate(self.objs):
+            if obj[0] == name:
+                return index, obj
+        return None, None