+++ CHANGELOG.md
@@ -13,4 +13,5 @@ added **randomColor3** + **randomColor4** + **genKey** + **saveData** + **loadDa
 
 # 25.06.2026
 
-added **checkInDebbuger** + **requirements.txt** + **builtin icons**
+updated **utils structure**
+added **checkInDebbuger** + **requirements.txt** + **builtin icons** + **loadSound** + **soundManager**